requirements.txt
Flask<1
mysql-connector-python<2